chore(i18n): move zh-tw catalogs to translation directory
