+++ src/main/services/__tests__/TopicNamingService.test.ts
@@ -75,7 +75,6 @@ const rendererI18nDir = path.join(process.cwd(), 'src/renderer/i18n')
 const unnamedTranslations = [
   'locales/en-us',
   'locales/zh-cn',
-  'locales/zh-tw',
   'translate/de-de',
   'translate/el-gr',
   'translate/es-es',
@@ -84,7 +83,8 @@ const unnamedTranslations = [
   'translate/pt-pt',
   'translate/ro-ro',
   'translate/ru-ru',
-  'translate/vi-vn'
+  'translate/vi-vn',
+  'translate/zh-tw'
 ].map((rel) => JSON.parse(fs.readFileSync(path.join(rendererI18nDir, `${rel}.json`), 'utf-8')).common.unnamed)
 
 function createService() {
