+++ src/renderer/components/history/__tests__/HistoryRecordsView.assistant.test.tsx
@@ -6,7 +6,6 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 import enUS from '../../../i18n/locales/en-us.json'
 import zhCN from '../../../i18n/locales/zh-cn.json'
-import zhTW from '../../../i18n/locales/zh-tw.json'
 import deDE from '../../../i18n/translate/de-de.json'
 import elGR from '../../../i18n/translate/el-gr.json'
 import esES from '../../../i18n/translate/es-es.json'
@@ -16,6 +15,7 @@ import ptPT from '../../../i18n/translate/pt-pt.json'
 import roRO from '../../../i18n/translate/ro-ro.json'
 import ruRU from '../../../i18n/translate/ru-ru.json'
 import viVN from '../../../i18n/translate/vi-vn.json'
+import zhTW from '../../../i18n/translate/zh-tw.json'
 
 const hookMocks = vi.hoisted(() => ({
   deleteTopic: vi.fn(),
