+++ src/renderer/i18n/resolver.ts
@@ -27,7 +27,7 @@ const logger = loggerService.withContext('I18N')
 const localeLoaders = {
   'en-US': () => import('./locales/en-us.json'),
   'zh-CN': () => import('./locales/zh-cn.json'),
-  'zh-TW': () => import('./locales/zh-tw.json'),
+  'zh-TW': () => import('./translate/zh-tw.json'),
   'de-DE': () => import('./translate/de-de.json'),
   'el-GR': () => import('./translate/el-gr.json'),
   'es-ES': () => import('./translate/es-es.json'),
